Update hooks.js: Fix outer page scrolling bug
Previously, the file used "scrollIntoView()" on the latest message element. This produced buggy behavior: When the chat was on a page that itself was scrollable, the entire page would scroll to get the bottom of the chat into view.

Currently, we 'simply' scroll the scrollable element of the chatbot all the way down. No matter where the bot is on the parent page, it will not influence the location of the parent page.

TODO: To get the exact same smooth scrolling behaviour, we need to add 
.simplebar-content-wrapper{scroll-behavior: smooth}