fix(react): don't copy multichannel props to created Text
to avoid messages:
ERROR Warning: React does not recognize the `firstIndex` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `firstindex` instead. If you accidentally passed it from a parent component, remove it from the DOM element. in message in Message in Text in Conte