Flow builder refactor getContents, handoff and action  (#2536)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
Refactor the getContents function to be able to get the contents in
several ways (through the node, Id or code), refactor the handoff logic
creating the FlowHandoff class. Also refactor the action to apply the
changes in the handoff and make it more readable.
<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context
Have several ways to get the contents through an id or the code. This
way we have clearer functions to obtain the contents through an Id in
the actions that are inside the bot and through the code for the
campaign plugin.
<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design
Refactor several things. 
The function getContents is now called getContentsByNode and only
accepts one node, you can no longer pass an id. The return is no longer
an object, it now returns the array of contents (the handoffNode
disappears and will be treated as a normal content).
I have created the functions getContentsById and getContentsByCode that
receive an Id or a code and end up calling getContentsByNode in this way
we can obtain the contents in 3 ways with the Id, the code or the node.
For the campaigns plugin we will use the getContentsByCode.

Refactor all handoff logic.
I have created the FlowHandoff class just like we have for FlowText,
FlowCarousel etc. This class has a function fromHubtypeCMS that creates
an instance of FlowHandoff from the handoff node. It also has the logic
to get the queue, the onFinishPayload and do the handoff.

Refactor the logic into action.
I have created functions with clearer names to indicate which node is
obtained in each case getNodeByUserInput getFallbackNode.
When the handoffNode disappears I apply a filter before passing the
contents to the render to do the handoff and eliminate that node.
<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**