Trigger Button with notifications (#2552)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
Be able to show in the trigger button the number of received and unread
messages. I consider a message unread when I receive it and the webchat
is closed.

<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context
In local is not working well as the message sets the ack to 1 directly.
Display a notification with the number of unread messages. 
CustomTriggerButton
![Captura de Pantalla 2023-07-04 a las 13 59
39](https://github.com/hubtype/botonic/assets/36898236/3e5adf7c-cfe6-469d-a3fd-77928dba685b)
DefaultTriggerButton

![DefaultTriggerButton](https://github.com/hubtype/botonic/assets/36898236/bc8ae49c-ec42-4b04-ba07-adfb13060b66)

<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design
Each time the action of webchatReducer addMessageComponent is called if
the webchat is closed and the message ack is 0, I increment the state of
unreadMessages by 1. When I reset unreadMessages

The div displaying unread messages has className='trigger-notifications'
to be able to adjust the css
<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**