Flow Builder: Add a variable in a text from Flow Builder (#2559)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
The objective is to be able to add a variable in a text from Flow
Builder.
<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context
For example if you put {language} in the text this part of the text will
be replaced by the current language which will be stored in
extra_data.language. If the variable does not exist in the session it
will be left {language}.
It also performs the fallback function. If extra_data is not defined, it
is set to isFirstFallbackOption = true.
<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design
I have made a refactor to have the request in the cmsApi class so it is
possible to replace the variables in the FLowText.fromHubtypeCMS
function. This way in the action we already have the final text and we
don't have to do extra logic in the action.

replaceVariables is a static function of the FlowText class. This
function receives the extraData object which is where the variable to be
replaced must be defined. This variable can be an empty string. This
function returns the replaced text.

<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**