Botonic react: multichannel text use whatsapp button list (#2606)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
Until now, when a text with more than 3 buttons was sent to the
multichannel and the bot was executed in whatsapp, several messages were
created and each one of them had 3 buttons. From now on we will use a
WhatsApp button list.
<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context
If a text arrives with more than 3 whatsapp buttons a whatsapp button
list will be created. But in a whatsapp button list you can't use
buttons pointing to urls or webviews.
A prop is added to the Multichannel, in the same way that a string could
already be passed to split messages ("More Options" by default). Now you
can pass a prop to define the text of the WhatsappButtonList button
("Menu" by default).
<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design
A whatsapp button list will be created for each group of 4 to 10 buttons
that have a payload or a path.
If there is a group of less than 4 buttons left, a message text will be
created with the missing buttons.
If the initial text had buttons that point to urls an extra message will
be added with the urls (they are not buttons but can be clicked to go to
the website).
If the initial text had buttons with webviews a message will be added
with the urls that lead to the webview.
![Captura de pantalla 2023-08-17 a las 16 57
13](https://github.com/hubtype/botonic/assets/36898236/192beb26-22c5-4d7c-b982-5233fa092233)
<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example
If you pass a text component like this to the multichannel
![Captura de pantalla 2023-08-17 a las 16 51
44](https://github.com/hubtype/botonic/assets/36898236/180b487b-3d88-4d68-a4f9-cdd6898233de)
the result in whatsapp will look like this
<img width="724" alt="Captura de pantalla 2023-08-17 a las 16 53 04"
src="https://github.com/hubtype/botonic/assets/36898236/8f1853a9-3bbe-4298-9dad-81ac936d8d61">

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing
Update the multichannel-text.test.jsx to check that a WhatsappButtonList
is used when there are more than 3 buttons.