Botonic react: WebchatState and webchatContext with typescript (#2608)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
In this PR the files related to the WebchatState with its two recudecrs
(webchatReducer and the messageReducer) and the WebchatContext with the
hook useWebchat are passed to typescript.
Apply a refactor to create a hooks folder and have each hook in a
separate file. All are exported from hooks/inde.ts

Set the typescript version to 5.0.4 if we upgrade to 5.2 we have
problems with the lint and it is also necessary to use in tsconfig.json
"module": "NodeNext", "moduleResolution": "NodeNext". I will check it in
the future.
<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context
The intention is to move all the code from botonic-react to typescript. 
<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design
Use the types already defined in the index-types files to write in
typescript the above mentioned files
<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->