Knowledge base plugin and inject to flow builder plugin  (#2610)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
**@botonic/plugin-knowledge-base**
This new plugin allows you to connect to a previously uploaded dashboard
knowledge base and then get answers about this document.

**@botonic/plugin-flow-builder**
You can pass the function to get the response from the knowledge base
through the constructor. If in the fallback node you have activated the
knowledge base, and it gives a correct response. The flow builder plugin
will create a node to respond with the information obtained by the
knowledge base plus the associated followup.

<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context
**@botonic/plugin-knowledge-base**
This plugin has the logic to make the call to our backend api using the
authToken to get the response from a knowledge base. In the constructor
it is necessary to pass the id of the knowledge base and the host where
the call is made https:api.hubtype.com/ for production, you can also
pass the authToken if you are developing locally and a timeout (usually
10 seconds).
Although it is a plugin it doesn't do anything in the pre or post
function, it only has a getIaResponse function that allows you to get
the response from the knowledge base. This response is an object of the
following type
```typescript
interface KnowledgebaseResponse {
    ai: string
    hasKnowledge: boolean
    sources: string[]
}
```
The parameter ai will have the answer and the hasKnowledge indicates if
this answer is related to the documents uploaded to the dashboard,
sources are the names of the documents used to get the answer.


**@botonic/plugin-flow-builder**
In the flow builder plugin there is no direct dependency on the
knowledge base plugin. You pass the function through the builder and
plug everything back in there. You add the logic to use the knowledge
base just before responding with the fallback.

<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design

<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example
First of all you need to add these two plugins independently in the bot
where you want to use them.

config.ts
```typescript
const knowledgeBase: PluginKnowledgeBaseOptions = {
  knowledgeBaseId: '50976b51-111-22ec-9bef-8888d25f2c5e',
  host: 'https://api.hubtype.com/',
  authToken: '51cf4b3b8316a02123456df789ef01',  // Used locally,
}
const flowBuilder:  BotonicPluginFlowBuilderOptions = {
    flowUrl: isProduction(env)
      ? 'https://api.ent0.flowbuilder.prod.hubtype.com/flow/latest'
      : 'https://api.dev.flowbuilder.dev.hubtype.com/flow/draft',
    getLocale: (session: Session) => context(session as BotSession).locale,
    getAccessToken: () => 'f3f123ab12345cde6789f03628d7fe', // Used locally,
    getKnowledgeResponse: async (request: BotRequest) => {
      const knowladgeBasePlugin = request.plugins
        .BotonicPluginKnowledgeBases as BotonicPluginKnowledgeBase
      return await knowladgeBasePlugin.getIaResponse(request.session)
    },
  }

export const config = {
    knowledgeBase,
    flowBuilder
}
```


plugins.ts
```typescript
export const plugins: core.PluginConfig<any>[] = [
  {
    id: 'BotonicPluginFlowBuilder',
    resolve: require('@botonic/plugin-flow-builder'),
    options: config.flowBuilder,
  },
  {
    id: 'BotonicPluginKnowledgeBases',
    resolve: require('@botonic/plugin-knowledge-bases'),
    options: config.knowledgeBase,
  }
]
```
<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**