Plugin Flow Builder: track knowledge base event (#2618)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description
Plugin-flow-builder and plugin-knowledge-bases:
- Change the type of sources in the knowledge base inference response
(it was not typed correctly). The attribute sources is an array of
objects, not a string array as it was until now.

Plugin-flow-builder:
- Adds botAiKnowladgeBase event tracking when the bot response is
obtained from the inference with the knowledge base
- Refactor: a small problem with the suffix on payloads. When adding a
suffix to all payloads when creating a button in flow builder, in the
routes we don't always have the expected payload. To solve this I remove
this suffix in the pre function of the plugin, this way in the routes we
have the "clean" payload.
<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->

## Context

<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->

## Approach taken / Explain the design

<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**

---------

Co-authored-by: Àlex Giménez Saludes <alex.gimenez.saludes@gmail.com>
Co-authored-by: Albert Gomáriz Sancha <64781562+AlbertGom@users.noreply.github.com>