Contentful Plugin: add content models with keywords to plugin options  (#2613)
<!-- _Set as [Draft
PR](https://github.blog/2019-02-14-introducing-draft-pull-requests/) if
it's not ready to be merged_.

[PR best practices
Reference](https://blog.codeminer42.com/on-writing-a-great-pull-request-37c60ce6f31d/)
-->

## Description

<!--
- Must be clear and concise (2-3 lines).
- Don't make reviewers think. The description should explain what has
been implemented or what it's used for. If a pull request is not
descriptive, people will be lazy or not willing to spend much time on
it.
- Be explicit with the names (don't abbreviate and don't use acronyms
that can lead to misleading understanding).
- If you consider it appropriate, include the steps to try the new
features.
-->
- Add `contentModelsWithKeywords` to `ContentfulOptions` to indicate
which content types has keywords
- Publish version `0.22.0`

## Context

<!--
- What problem is trying to solve this pull request?
- What are the reasons or business goals of this implementation?
- Can I provide visual resources or links to understand better the
situation?
-->
We needed to add keywords to the handoff content to just one bot. So we
needed to define the contents with keywords depending on the bot.

## Approach taken / Explain the design

<!--
- Explain what the code does.
- If it's a complex solution, try to provide a sketch.
-->

## To document / Usage example

<!--
- How this is used?
- If possible, provide a snippet of code with a usage example.
-->

## Testing

The pull request has one unit test checking that
`contentModelsWithKeywords` option is being applied correctly.

---------

Co-authored-by: Adrià Sastre <adria.sastre@gmail.com>