Contentful plugin publish 0.23.1 (#2668)
## Description
In the free Contentful account where we have everything related to
testing, now we can only have master and 2 environments. We had a test
that creates and deletes an environment. This test was failing because
there was not enough space. I have merged 2 environments into one
(deleting the testx environment after copying all the content to
manage-contentful.
Released plugin version 0.23.1