Feature: Unread messages (#2626)
## Description
This feature consists of counting the messages not read by the user and
displaying a banner (UnreadMessageBanner) above the last unread message
indicating how many unread messages there are. This number is the same
as the number shown on the Trigger Button.
This feature also disables the automatic scrolling to read from the last
unread message and if you are not at the bottom of the scroll it shows a
button (ScrollButton) at the bottom left that scrolls to the bottom when
clicked.
Both the UnreadMessagesBanner and the ScrollButton can be Custom
Components that are passed as a custom property in the webchat theme.

## Context
In this image you can see the UnreadMessagesBanner and the ScrollButton

![Captura de pantalla 2023-10-04 a las 16 52
54](https://github.com/hubtype/botonic/assets/36898236/f47363f3-3320-4db4-8ec1-9e1ddf858b0f)

## Approach taken / Explain the design
To review this PR better to group by commits

- The first commit disables scrolling at all points.

- The second commit creates the UnreadMessagesBanner and ScrollButton
components.

- The third commit allows to pass custom properties to the webchat Theme
to activate these components or to pass Custom Components to replace
these components. If you pass a custom component it is not necessary to
activate each of these components.If you pass notifications: { enable:
true } both components are enabled.
We had a custom property in the Theme that was used to activate this
number of unread messages in the TriggerButton but from now on this will
also be activated with the notification: { triggerButton: { enable: true
}} (TODO: maybe it should also be activated by default when activating
notifications in general).

- The fourth commit only adds the images used in the
UnreadMessagesBanner and ScrollButton components.

- The fifth commit adds the functions and status in the reducer for
resetUnreadMessages, setLastMessageVisible which are used later when
rendering the whole list of messages.

- The sixth commit replaces the props.from of the message with the
sentBy that is obtained from the backend. This is used to only count
non-user messages as unread. The prop.sentBy will allow us in a future
PR to differentiate the image that is shown in a message sent by the bot
and agent.

- The seventh commit creates the MessageList component which is in
charge of rendering all messages and adding the UnreadMessagesBanner on
top of the last unread message and the ScrollButton when necessary.

- The eighth commit I have applied feedback from the PR: rename
unreadMessages to numUnreadMessages, simplify renders and code, typo
errors

- The ninth commit I added the automatic scroll to center in the screen
UnreadMesssagesBanner when you open the webchat (and the banner exists)

- The tenth commit is the WIP with things that I want to discuss

## To document / Usage example
Possible ways to activate this feature from the webchat custom
properties in src/webchat/index.ts

Activate with all default components
```typescript
export const webchat: WebchatArgs = {
  ...
  them: {
    ...
    notifications: { enable: true, triggerButton: { enable: true } },
    scrollButton: { enable: true },
    ...
}
```

Activate using custom components
```typescript
import CustomScrollButton, CustomUnreadMessagesBanner from './customs'

export const webchat: WebchatArgs = {
  ...
  them: {
    ...
    notifications: { custom: CustomUnreadMessagesBanner, triggerButton: { enable: true } },
    scrollButton: { custom: CustomScrollButton },
    ...
    }
```


## Testing
To fix the tests I created another PR because I needed to modify a lot
of files.

---------

Co-authored-by: Adrià Sastre <adria.sastre@gmail.com>