Bot image or Agent image on MessageTimestamp component  (#2670)
## Description
Add in the ThemeProps message.timestamp.withImage to render next to the
timestamp the image of the bot or the agent depending on who sends the
message. If the agent's image is not defined it will always display the
bot's image.

## Context
To be able to make this message design easier to use.
![Captura de pantalla 2023-11-21 a las 18 36
07](https://github.com/hubtype/botonic/assets/36898236/8fe68da6-4335-476b-951f-753b6ecda6de)

## To document / Usage example
In the code of the `webchat/index.ts` bot you can define the images and
have them rendered in the timestamp component.


```typescript
export const webchat: WebchatArgs = {
...
   message: {
   ...
      bot: {
        style: {},
        image: BotImage,
      },
      agent: {
        image: AgentImage,
      },
      timestamps: {
        withImage: true,
        format: () => createFormatTimestamp(),
      },
    },
   ...
```

It is no longer necessary to put css to edit this component. But if you
want to edit the css of the MessageTimestamp you now have two empty css
classes:
- botonic-timestamp-container 
- bot or agent