Plugin-flow-builder: remove start_up_node (#2686)
## Description
- Remove the start_up_node. Always use the start_node_id to retrieve the
content used as start node.
- Add a getStartContents function

## Context
Until now we had one node (different from all) for which text content
was the start_up_node. With this change the JSON that has all the flow
has an attribute where the id of the content that is used as start is
indicated (it doesn't have to be a text content).