Plugin flow builder: open user input in webchat when do handoff (#2694)
## Description
When doing a handoff you need to have the userInput of the webchat
enable and this is not always the case for all bots.
From the flow builder plugin always enable userInput when do a handoff.
If no other message is sent with the WebchatSettings the typing is set
to true. That's why I also modify @botonic/react to set it to false when
updating the WebchatSettings.